Adapted to new rolling upgrade protocol
where the first slave to do the upgrade does a store migration, but the
other slaves remove their local stores and grabs a backup from that slave
that did the store migration instead of doing the migration on their own.
This is so since there's no other known way of deterministically ending up
with the same store layout, specifically around the assignment of new ids
for relationship group records.