Integrate Muninn
And make it the default PageCache implementation in Enterprise.

Also fix an issue where the NeoStore flushed its PagedFile while holding an exclusive lock on a page.
This scenario needs to be properly specified.
It must either throw and exception when attempted, or be explicitly allowed.
Whatever the case, it's not something that should differ between the StandardPageCache and the MuninnPageCache implementations.