Changes store application synchronization
this opens the flood gates for potential problems, unexpected problems
that we want to see at this stage.

This commit also introduces synchronization on label scan store and schema
indexing application, something that the (above) removed synchronization
actually didn't include, hence the cause of recent problems seen in label
scan store and schema indexes in 2.2.