No longer closes txId on append failure
since we now distunguish between committing and committed txId. The
appender gets a next committing txId, which is just a candidate for
committing. So in case of failure to append there's no transaction to
close since it's not marked as committed. Marking as committed is the last
operation of appending.

This issue would be observed by having a transaction failing to append to
the log, to the go and shut down the database. The database would await
last closed transaction to be the same as the last committed, and since
last closed would be lastCommitted+1 that would never happen.