String->long in-memory efficient index for batch importer
for use in cases where the node ids given in the node/relationship input
aren't actual record ids, but arbitrary strings. When using this type of
IdMapping the node ids (strings) are fed to the in-memory index during the
node import stage and sorted, and general prepared for getting
String->long later in other stages that deals with relationships where
they also specify these strings as the start/end node ids.