Exact number of relationships given node label.
The CountsStore maintains the exact number of relationships in the
database that either starts or ends in a node with a given label. Counts
are maintained both in total, and by relationship type, but only for one
label, i.e. number of relationships that starts in a node with a given
label and ends in a node with a given label is not maintained.